fix: two bean of s3service
