+++ src/main/java/school/hei/haapi/service/TranscriptVersionService.java
@@ -6,21 +6,9 @@ import org.springframework.stereotype.Service;
 import school.hei.haapi.model.Transcript;
 import school.hei.haapi.model.TranscriptVersion;
 import school.hei.haapi.model.User;
-import school.hei.haapi.model.exception.NotFoundException;
 import school.hei.haapi.repository.TranscriptRepository;
 import school.hei.haapi.repository.TranscriptVersionRepository;
 import school.hei.haapi.repository.UserRepository;
-import school.hei.haapi.service.aws.S3Service;
-import school.hei.haapi.service.utils.FilePdfUtils;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-import school.hei.haapi.model.TranscriptVersion;
-import school.hei.haapi.repository.TranscriptVersionRepository;
 
 @Service
 @AllArgsConstructor
